+++ app/assets/javascripts/application.js
@@ -1,3 +1,3 @@
 //= require_tree ./modules
 //= require govuk_publishing_components/components/error-summary
-//= require govuk_publishing_components/components/task-list
+//= require govuk_publishing_components/components/step-by-step-nav
