+++ app/assets/stylesheets/print.scss
@@ -3,6 +3,6 @@
 @import 'typography';
 
 @import 'components/print/*';
-@import 'govuk_publishing_components/components/print/task-list';
-@import 'govuk_publishing_components/components/print/task-list-header';
+@import 'govuk_publishing_components/components/print/step-by-step-nav';
+@import 'govuk_publishing_components/components/print/step-by-step-nav-header';
 @import 'print/html-publication';
