+++ app/views/content_items/answer.html.erb
@@ -1,5 +1,1 @@
-<% if current_tasklist_ab_test.show_tasklist_sidebar? %>
-  <%= render 'content_items/body_with_related_links', tasklist_content: tasklist_content %>
-<% else %>
-  <%= render 'content_items/body_with_related_links' %>
-<% end %>
+<%= render 'content_items/body_with_related_links' %>
