+++ app/views/content_items/publication.html.erb
@@ -1,4 +1,4 @@
-<%= content_for :no_breadcrumbs, current_tasklist_ab_test.show_tasklist_header? %>
+<%= content_for :no_breadcrumbs, show_step_nav? %>
 <%= content_for :simple_header, @navigation.should_present_taxonomy_navigation? %>
 
 <div class="grid-row">
