+++ app/views/layouts/application.html.erb
@@ -25,11 +25,6 @@
   <% end %>
   <meta name="govuk:content-id" content="<%= @content_item.content_id %>" />
 
-  <% if current_tasklist_ab_test.eligible? %>
-    <%= current_tasklist_ab_test.sidebar_variant.analytics_meta_tag.html_safe %>
-    <%= current_tasklist_ab_test.header_variant.analytics_meta_tag.html_safe %>
-  <% end %>
-
   <%= yield :extra_head_content %>
 </head>
 <body>
@@ -39,8 +34,8 @@
 
   <div id="wrapper" class="<%= wrapper_class %>">
     <%= render_phase_label @content_item, content_for(:phase_message) %>
-    <% if current_tasklist_ab_test.show_tasklist_header? && local_assigns.has_key?(:tasklist_content) %>
-      <%= render 'shared/tasklist_header', tasklist_content: local_assigns[:tasklist_content] %>
+    <% if show_step_nav?%>
+      <%= render 'shared/step_nav_header', step_nav_content: current_step_nav %>
     <% else %>
       <% if !content_for(:no_breadcrumbs) %>
         <%= render 'shared/breadcrumbs' %>
