--- app/views/shared/_sidebar_tasklist.html.erb
@@ -1,10 +1,0 @@
-<%
-  no_margin ||= false
-%>
-<nav class="column-third <%= "task-list-sidebar" unless no_margin %>">
-  <%= render 'govuk_publishing_components/components/task_list_related', links: [{
-      href: tasklist_content.base_path,
-      text: tasklist_content.title
-  }] %>
-  <%= render 'govuk_publishing_components/components/task_list', tasklist_content.set_current_task %>
-</nav>
