+++ app/views/shared/_step_nav_header.html.erb
@@ -1,0 +1,5 @@
+<%= render "govuk_publishing_components/components/step_by_step_nav_header",
+  title: current_step_nav.title,
+  path: current_step_nav.base_path,
+  skip_link: current_step_nav.skip_link
+  %>
