--- app/views/shared/_tasklist_header.html.erb
@@ -1,5 +1,0 @@
-<%= render "govuk_publishing_components/components/task_list_header",
-  title: tasklist_content.title,
-  path: tasklist_content.base_path,
-  skip_link: tasklist_content.skip_link
-  %>
